Relationship weight double

+++ src/edu/isistan/christian/recommenders/groups/magres/MAGReS.java
@@ -45,7 +45,7 @@ public abstract class MAGReS <T extends SURItem> extends GroupRecommender<T>{
 	 * All the agents will use the default profile. This method exists just to comply with the interface
 	 * of the GroupRecommender class
 	 */
-	public GRecResult<T> recommend(GRecGroup group, int howMany, HashMap<SURUser, Double> assertivenessFactors, HashMap<SURUser, Double> cooperativenessFactors, HashMap<SURUser, HashMap<SURUser, Integer>> relationshipsFactors) throws SURException{
+	public GRecResult<T> recommend(GRecGroup group, int howMany, HashMap<SURUser, Double> assertivenessFactors, HashMap<SURUser, Double> cooperativenessFactors, HashMap<SURUser, HashMap<SURUser, Double>> relationshipsFactors) throws SURException{
 		return this.recommend(group, null, howMany, assertivenessFactors, cooperativenessFactors, relationshipsFactors);
 	}
 	
@@ -57,7 +57,7 @@ public abstract class MAGReS <T extends SURItem> extends GroupRecommender<T>{
 	 * @return
 	 * @throws SURException
 	 */
-	public abstract GRecResult<T> recommend (GRecGroup group, HashMap<SURUser, PUMASAgentProfile<T>> userAgProfiles, int howMany, HashMap<SURUser, Double> assertivenessFactors, HashMap<SURUser, Double> cooperativenessFactors, HashMap<SURUser, HashMap<SURUser, Integer>> relationshipsFactors) 
+	public abstract GRecResult<T> recommend (GRecGroup group, HashMap<SURUser, PUMASAgentProfile<T>> userAgProfiles, int howMany, HashMap<SURUser, Double> assertivenessFactors, HashMap<SURUser, Double> cooperativenessFactors, HashMap<SURUser, HashMap<SURUser, Double>> relationshipsFactors) 
 			throws SURException;
 
 }
